telemetry-logger/code/requirements.txt
influxdb-client[ciso]
python-kasa
requests